modeljson/function-mi-akv-sql/requirements.txt
azure-functions
adal
msrestazure
pyodbc
azure-keyvault==1.1.0